pipeline/requirements.txt
-e git+ssh://git@github.com/SoheilKhodayari/OpenRedirect.git@9b6a876f70bc50ae405d8614326f3b9661e5a196#egg=candidate_collector&subdirectory=code/candidate_collector
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
docutils==0.19
-e git+ssh://git@github.com/SoheilKhodayari/OpenRedirect.git@9b6a876f70bc50ae405d8614326f3b9661e5a196#egg=endpoint_analyzer&subdirectory=code/endpoint_analyzer
greenlet==2.0.1
idna==3.4
lockfile==0.12.2
luigi==3.1.1
-e git+ssh://git@github.com/SoheilKhodayari/OpenRedirect.git@9b6a876f70bc50ae405d8614326f3b9661e5a196#egg=or_scanner&subdirectory=code/or_scanner
playwright==1.28.0
pyee==9.0.4
python-daemon==2.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
tenacity==8.1.0
tornado==6.2
typing_extensions==4.4.0
urllib3==1.26.13
